requirements.txt
sentence-transformers
faiss-cpu
google-generativeai
groq

